dev-requirements.txt
-r requirements.txt

django-pdb
django-extensions
django-debug-toolbar